Only tweaks

+++ removeOffer.php
@@ -1,9 +1,9 @@
 <?php
 include_once './core/session.php';
-if (!empty($_SERVER['HTTP_X_REQUESTED_WITH']) && strtolower($_SERVER['HTTP_X_REQUESTED_WITH']) == 'xmlhttprequest') {
+if (!empty($_SERVER['HTTP_X_REQUESTED_WITH']) && strtolower($_SERVER['HTTP_X_REQUESTED_WITH']) == 'xmlhttprequest' && !empty($user["id"])) {
     if ($_POST) {
         $offer = (int) cleanString($_POST["id"]);
-        if (!empty($offer)) {
+        if (!empty($offer) && my_offer($offer, $user["id"])) {
             if (Db::query("DELETE FROM cart WHERE id = ? LIMIT 1", $offer) == 1) {
                 echo "success|Del uspešno odstranejen iz košarice!";
             } else {
